+++ lib/Adafruit_SSD1351-gemu-1.0/Tiger.rgb
