+++ lib/JaretBurkett_ILI9488-gemu-1.0/library.properties
